Refactor backend app routes -part1
